abstract the community tabs header
